replace selectors
