+++ packages/eas-cli/src/rollout/actions/SelectRollout.ts
@@ -1,6 +1,8 @@
 import { SelectChannel } from '../../channel/actions/SelectChannel';
 import { EASUpdateAction, EASUpdateContext } from '../../eas-update/utils';
 import { UpdateChannelBasicInfoFragment } from '../../graphql/generated';
+import { Ora, ora } from '../../ora';
+import { Edge, QueryParams } from '../../utils/relay';
 import { isRollout } from '../branch-mapping';
 
 /**
@@ -8,10 +10,29 @@ import { isRollout } from '../branch-mapping';
  */
 export class SelectRollout implements EASUpdateAction<UpdateChannelBasicInfoFragment | null> {
   public async runAsync(ctx: EASUpdateContext): Promise<UpdateChannelBasicInfoFragment | null> {
-    const selectChannel = new SelectChannel({
+    let assetSpinner: Ora | null = null;
+    const afterEachFilterQuery = (
+      _externalQueryParams: QueryParams,
+      totalNodesFetched: number,
+      _dataset: Edge<UpdateChannelBasicInfoFragment>[],
+      willFetchAgain: boolean
+    ): void => {
+      if (willFetchAgain && !assetSpinner) {
+        assetSpinner = ora().start('Fetching channels...');
+      }
+      if (assetSpinner) {
+        assetSpinner.text = `Fetched ${totalNodesFetched} channels`;
+      }
+    };
+    const selectChannelAction = new SelectChannel({
       printedType: 'rollout',
       filterPredicate: isRollout,
+      afterEachFilterQuery,
     });
-    return await selectChannel.runAsync(ctx);
+    const channelInfo = await selectChannelAction.runAsync(ctx);
+    if (assetSpinner) {
+      (assetSpinner as Ora).succeed(`Fetched all channels`);
+    }
+    return channelInfo;
   }
 }
